chore: complete Ts.ED refactoring Phase 6 - testing cleanup
Complete the final phase of Ts.ED framework refactoring by:
- Creating domain test files for pure domain logic (AgentTools, PlanFormatter, WorkspaceManager)
- Deleting 8 legacy test files replaced by new Ts.ED-organized structure
- Updating jest.config.js to exclude legacy tests from execution

This completes the migration to a clean Ts.ED architecture with proper separation:
- tests/controllers/ - HTTP endpoint tests
- tests/services/ - DI service tests
- tests/domain/ - Pure domain logic tests (no framework decorators)
- tests/middlewares/ - Middleware tests
- tests/integration/ - End-to-end tests

All 213 tests pass with the new structure.

Related to Ts.ED refactoring phases 0-6 documented in TSED_REFACTOR_STRATEGY.md
