+++ jest.config.js
@@ -4,11 +4,14 @@ module.exports = {
   testMatch: ['**/tests/**/*.test.ts'],
   testPathIgnorePatterns: [
     '/node_modules/',
-    'tests/server.test.ts',      // Excluded - needs PlatformTest rewrite (task 020)
-    'tests/agent.test.ts',       // Excluded - needs service class rewrite (task 021)
-    'tests/worker.test.ts',      // Excluded - needs service class rewrite (task 021)
-    'tests/plan-store.test.ts',  // Excluded - needs service class rewrite (task 021)
-    'tests/linear-client.test.ts' // Excluded - needs service class rewrite (task 021)
+    'tests/server.test.ts',        // Excluded - replaced by tests/controllers/ + tests/integration/
+    'tests/agent.test.ts',         // Excluded - replaced by tests/services/AgentOrchestratorService.test.ts
+    'tests/worker.test.ts',        // Excluded - replaced by tests/services/WorkerService.test.ts
+    'tests/plan-store.test.ts',    // Excluded - replaced by tests/services/PlanStoreService.test.ts
+    'tests/linear-client.test.ts', // Excluded - replaced by tests/services/LinearClientService.test.ts
+    'tests/tools.test.ts',         // Excluded - replaced by tests/domain/AgentTools.test.ts
+    'tests/workspace.test.ts',     // Excluded - replaced by tests/domain/WorkspaceManager.test.ts
+    'tests/plan-formatter.test.ts' // Excluded - replaced by tests/domain/PlanFormatter.test.ts
   ],
   collectCoverage: true,
   collectCoverageFrom: [
